remove state from login components
