--- src/components/LoginForm/index.ts
@@ -1,3 +1,0 @@
-import LoginForm from "./LoginForm";
-
-export default LoginForm;
