--- src/components/PlayerCard/index.ts
@@ -1,3 +1,0 @@
-import PlayerCard from "./PlayerCard";
-
-export default PlayerCard;
