+++ src/containers/GamePageContainer/GamePageContainer.tsx
@@ -2,7 +2,6 @@ import React from "react";
 import { getCurrentUserSync } from "@/utils/AuthUtils";
 import { GameSettings } from "@/types/GameTypes";
 import { GameMode } from "@/enums/GameEnums";
-import PlayerCard from "@/components/PlayerCard";
 import GameForm from "@/components/GameForm";
 import GameAreaContainer from "@/containers/GameAreaContainer";
 
@@ -42,7 +41,6 @@ class GamePageContainer extends React.PureComponent<unknown, GamePageContainerSt
   render(): JSX.Element {
     return (
       <>
-        <PlayerCard userName={this.state.userName} />
         <GameForm initialValues={this.state.settings} onSubmit={this.handleSubmit} onReset={this.handleReset} />
         {this.state.isStarted && <GameAreaContainer settings={this.state.settings} />}
       </>
