+++ src/modules/Authentication/index.ts
@@ -1,3 +1,4 @@
 export { authenticationSlice } from "./slice";
 export { authenticationSaga } from "./saga";
 export { AuthRoute } from "./Route/AuthRoute";
+export { LoginForm } from "./Form/LoginForm";
