+++ src/screens/GameScreen.tsx
@@ -1,0 +1,4 @@
+import React from "react";
+import { PlayerCard } from "@/modules/Authentication/PlayerCard/PlayerCard";
+
+export const GameScreen = (): JSX.Element => <PlayerCard />;
