+++ src/screens/LoginScreen.tsx
@@ -1,0 +1,4 @@
+import React from "react";
+import { LoginForm } from "@/modules/Authentication";
+
+export const LoginScreen = (): JSX.Element => <LoginForm />;
