+++ src/screens/index.ts
@@ -1,0 +1,2 @@
+export { LoginScreen } from "./LoginScreen";
+export { GameScreen } from "./GameScreen";
